% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{choc_ctest}
\alias{choc_ctest}
\title{Collapse Test on different chocolate}
\format{A data frame with 240 rows and 3 variable:
	\describe{
			\item{Test}{factor, names of the three differen chocolate}
			\item{Temperature}{numeric, temperature of analysis}
			\item{Gap}{numeric, height of the sample during the analysis}
	}}
\usage{
choc_ctest
}
\description{
A dataset containing the results of a series of Collapse Test analysis
	on three different chocolate:
	Ferrero latte 2001
	Barry   Thorntons standard chocolate (containing 7% of crumb)
	Barry   Thorntons eritage chocolate (containing 24% of crumb)
}
\keyword{datasets}

