% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{composition}
\alias{composition}
\title{Composition of Thorntons Continental}
\format{A data frame 1 rows and 10 variables:
	\describe{
			\item{Alpini}{numbers of Alpine praline present in the box}
			\item{VanillaT}{numbers of Vanilla Truffle praline present in the box}
			\item{Sicilian}{numbers of Sicilian Lemon praline present in the box}
			\item{Apricot}{numbers of Apricot Danish praline present in the box}
			\item{Coffee}{numbers of Coffee praline present in the box}
			\item{Hslice}{numbers of Hazelnut Slice Lemon praline present in the box}
			\item{Milano}{numbers of Milano praline present in the box}
			\item{Mix}{numbers of Mix praline present in the box}
			\item{Orange}{numbers of Orange praline present in the box}
			\item{Salted}{numbers of Salted Caramel praline present in the box}
	}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
composition
}
\description{
A dataset containing the number of each praline present in the box
	the column name indicate the type of praline present in the box
	in the first row is inserted the numebr of that type present
	in the 'source.rda' file there are the weight of a sample of the
	same praline
}
\keyword{datasets}

