% Generated by roxygen2 (4.0.2): do not edit by hand
\name{contraction}
\alias{contraction}
\title{contraction}
\usage{
contraction(path = "~/SFTP_Area/ITA18_Project_Area/PTPM_Praline", nome = "",
  poly.gr = 6, d1_peak = 2, d2_peak = c(3, 4))
}
\arguments{
\item{path}{file path}

\item{nome}{file name}

\item{poly.gr}{polynomial degree}

\item{d1_peak}{first derivative peak number}

\item{d2_peak}{second derivative peak numbers}
}
\value{
list results 'dati', 'header'
}
\description{
ChocoAnalyzer contraction elaboration
}
\examples{
#
}

