% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ctest}
\alias{ctest}
\title{collapse test}
\usage{
ctest(analisi, xlim, ylim, main = "Collapse test", point = F, values = F,
  dt = F, mm = 1)
}
\arguments{
\item{analisi}{data frame of data [Test, Temperature, Gap]}

\item{xlim}{x axis limits - default c(26, 35)}

\item{ylim}{y axis limits - default c(0, 5)}

\item{main}{graph title - default 'Collapse Test}

\item{point}{boolean T = plot points of d' and d'' - default F}

\item{values}{boolean T = plot values of d' and d'' - default F}

\item{dt}{boolean T = plot derivatives - default F}

\item{mm}{integer media mobile se >1}
}
\value{
analysis list (calculated data and results)
}
\description{
Calculate the melting point of a chocolate by collapse test analysis
}
\examples{
#data(choc_ctest)
#ctest(choc_ctest)
}

