% Generated by roxygen2 (4.0.2): do not edit by hand
\name{derive}
\alias{derive}
\title{derive}
\usage{
derive(x, y, step = 3)
}
\arguments{
\item{x}{x vector}

\item{y}{y vector}

\item{step}{value to be interpoleted for calculation (3 or 5)}
}
\value{
gradient vector
}
\description{
Calculate gradient y dx
}
\examples{
#data(choc_ctest)
#ctest(choc_ctest)
}

