% Generated by roxygen2 (4.0.2): do not edit by hand
\name{enrobing}
\alias{enrobing}
\title{enrobing}
\usage{
enrobing(input = 1000, output = 300, ricircolo = 2000, vasca = 120,
  minuti = 600, xlim = NULL, ylim = NULL, debug = F, gg = F)
}
\arguments{
\item{input}{vector of flow of fresh chocolate from tempering in kg/h - defuault = 1000 kg/h}

\item{output}{vector of chocolate consumption in kg/h - default = 300/kg/h}

\item{ricircolo}{vector of flow of reclicling pump in kg/h - default = 2000 kg/h}

\item{vasca}{vector of capacity of enrobing tub in kg - default = 120 kg}

\item{minuti}{minute of simulation - default = 600}

\item{xlim}{x axes extension - defaul = Null (automatic)}

\item{ylim}{y axes extension - defaul = Null (automatic)}

\item{debug}{if T debug is on - default = F}

\item{gg}{if T use ggplot2 if F use standard graphics - default = F}
}
\value{
dataframe (step, composition, percentage)
}
\description{
Simulate enrobing indirect system
}
\examples{
# enrobing()
}

