% Generated by roxygen2 (4.0.2): do not edit by hand
\name{err_ce}
\alias{err_ce}
\title{err_ce}
\usage{
err_ce(decweight)
}
\arguments{
\item{decweight}{the declared weight}
}
\value{
error
}
\description{
Compute tollerance of the weight
}
\details{
this is detail field bla bla bla
  bla bla
}
\examples{
#err_ce(200)
}

