% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{L1_max}
\alias{L1_max}
\title{Weight of single Rocher praline}
\format{A data frame with 450 rows and 1 variable:
	\describe{
			\item{RO}{weight of 450 Rocher}
	}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
L1_max
}
\description{
A dataset containing the weight of the single Rocher praline
	produced the 11 May 2016 on Linea 1 of Alba
	The date of the 11th of May was the data where to Linea 1
	has had the maximum weigth of the year 12.726g, sd=0.266
}
\keyword{datasets}

