% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{L1_min}
\alias{L1_min}
\title{Weight of single Rocher praline}
\format{A data frame with 939 rows and 1 variable:
	\describe{
			\item{RO}{weight of 939 Rocher}
	}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
L1_min
}
\description{
A dataset containing the weight of the single Rocher praline
	produced the 11 November 2016 on Linea 1 of Alba
	The date of the 11th of November was the data where to Linea 1
	has had the minimun weigth of the year 12.695g, sd=0.287
}
\keyword{datasets}

