% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{L4_max}
\alias{L4_max}
\title{Weight of single Rocher praline}
\format{A data frame with 310 rows and 1 variable:
	\describe{
			\item{RO}{weight of 310 Rocher}
	}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
L4_max
}
\description{
A dataset containing the weight of the single Rocher praline
	produced the 22 October 2016 on Linea 4 of Alba
	The date of the 22nd of October was the data where to Linea 4
	has had the maximum weigth of the year 12.840g, sd=0.292
}
\keyword{datasets}

