% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{L4_min}
\alias{L4_min}
\title{Weight of single Rocher praline}
\format{A data frame with 479 rows and 1 variable:
	\describe{
			\item{RO}{weight of 479 Rocher}
	}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
L4_min
}
\description{
A dataset containing the weight of the single Rocher praline
	produced the 23 September 2016 on Linea 4 of Alba
	The date of the 23rd of September was the data where to Linea 4
	has had the minimun weigth of the year 12.705g, sd=0.295
}
\keyword{datasets}

