% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mosaicFB}
\alias{mosaicFB}
\title{mosaicFB}
\usage{
mosaicFB(dati, cluster = "plant", filter = NULL, depivot = T,
  main = "Product fat-bloom", sub = "")
}
\arguments{
\item{cluster}{field to use to clusterize data}

\item{filter}{data filter}

\item{depivot}{boolean, if T depivot data, if F none}

\item{main}{main title}

\item{sub}{subtitle}

\item{data}{data frame of data}
}
\value{
elaborated data frame
}
\description{
Simulate chocolate crystallization using Imogen Foubert model
  of crystallization kinetics of cocoa butter
}
\examples{
#tempering(BC=0.42)
}

