% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pa2pa}
\alias{pa2pa}
\title{pa2pa}
\usage{
pa2pa(na1, pz1, na2, pz2, xlim = c(0, 0.1), alfa = 0.05)
}
\arguments{
\item{na1}{numero di difetti accettati piano di campionamento 1}

\item{pz1}{numero di pezzi campionati piano di campionamento 1}

\item{na2}{}

\item{pz2}{}

\item{xlim}{scala dell'asse x}

\item{alfa}{errore alfa ammesso}
}
\value{
none
}
\description{
Compare two attributive sample plan
}
\examples{
# pa2pa(0, 125, 3, 125)
}

