% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{RO_ALL}
\alias{RO_ALL}
\title{Temperindex analysis on stored Rocher of all plant}
\format{A data frame with 234 rows and 15 variable:
	\describe{
			\item{plant}{factor, name of ptest TI5=product with TI=5, TI7= product with TI=7}
			\item{lotto}{factor, }
			\item{data-produzione}{factor, }
			\item{data_reale}{factor, }
			\item{eta_reale}{int, }
			\item{mese}{factor, months of analysis after production}
			\item{confezione}{int, }
			\item{linea}{factor, }
			\item{L1}{numeric, number of product with level 1 of fat bloom}
			\item{L2}{numeric, number of product with level 2 of fat bloom}
			\item{L3}{numeric, number of product with level 3 of fat bloom}
			\item{L4}{numeric, number of product with level 4 of fat bloom}
			\item{L5}{numeric, number of product with level 5 of fat bloom}
			\item{L6}{numeric, number of product with level 6 of fat bloom}
	}}
\usage{
RO_ALL
}
\description{
A dataset containing the results of fat bloom analysis during 6 months
	with Ferrero Rocher of all production plant
	Ferrero latte 2001
}
\keyword{datasets}

