% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{RO_ALL_DE}
\alias{RO_ALL_DE}
\title{Temperindex analysis on stored Rocher of all plant (de-pivot data)}
\format{A data frame with 3694 rows and 9 variable:
	\describe{
			\item{plant}{factor, name of ptest TI5=product with TI=5, TI7= product with TI=7}
			\item{lotto}{factor, }
			\item{data-produzione}{factor, }
			\item{data_reale}{factor, }
			\item{eta_reale}{int, }
			\item{mese}{factor, months of analysis after production}
			\item{confezione}{int, }
			\item{linea}{factor, }
			\item{affioramento}{numeric, fat bloom level of the sample (single praline)}
	}}
\usage{
RO_ALL_DE
}
\description{
A dataset containing the results of fat bloom analysis during 6 months
	with Ferrero Rocher of all production plant
	Ferrero latte 2001
}
\keyword{datasets}

