% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{RO_OHG}
\alias{RO_OHG}
\title{Temperindex on Stadtallendorf test}
\format{A data frame with 12 rows and 8 variable:
	\describe{
			\item{prodotto}{factor, name of ptest TI5=product with TI=5, TI7= product with TI=7}
			\item{mese}{factor, months of analysis after production}
			\item{L1}{numeric, number of product with level 1 of fat bloom}
			\item{L2}{numeric, number of product with level 2 of fat bloom}
			\item{L3}{numeric, number of product with level 3 of fat bloom}
			\item{L4}{numeric, number of product with level 4 of fat bloom}
			\item{L5}{numeric, number of product with level 5 of fat bloom}
			\item{L6}{numeric, number of product with level 6 of fat bloom}
	}}
\usage{
RO_OHG
}
\description{
A dataset containing the results of fat bloom analysis during 6 months
	with 2 different product cluster:
	TI5   Rocher produced with 2nd enrobing tempered with TI=5
	TI7   Rocher produced with 2nd enrobing tempered with TI=7
}
\keyword{datasets}

