% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simHollow}
\alias{simHollow}
\title{simHollow}
\usage{
simHollow(g = matrix(c(0, -1, 0), ncol = 1), rpm1 = 1, rpm2 = 6,
  rpm3 = 0, sim = 120, step = 2, seq = F, d3 = T)
}
\arguments{
\item{g}{define gravity vector}

\item{rpm1}{asse x - 0= no rotazione >0 rotazione oraria <0 rotazione antioraria}

\item{rpm2}{asse z - 0= no rotazione >0 rotazione oraria <0 rotazione antioraria}

\item{rpm3}{asse y - 0= no rotazione >0 rotazione oraria <0 rotazione antioraria}

\item{sim}{totale tempo simulato in secondi}

\item{step}{passo di simulazione in minuti}

\item{seq}{if True plot 'sequential' graph if False plot 'stellar' graph}

\item{d3}{if True plot 3d graph if False plot 2d graph}
}
\value{
coordinate
}
\description{
Simulate molds movements during hollow figure production
}
\examples{
#simHollow()
}

