% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulateMH}
\alias{simulateMH}
\title{simulateMH}
\usage{
simulateMH(source, target, declared, min.pcs, step = 1000)
}
\arguments{
\item{source}{vector with the weight of the single praline}

\item{target}{the target weight of the multi-head machine}

\item{declared}{declared weight of the box (unused)}

\item{min.pcs}{declared pieces of the box (unused)}

\item{step}{simulation steps, number of pack to be simulated}
}
\value{
data.frame width data simulation results
}
\description{
Simulate a multi-head weighting machine producing a box of pralines
the result of the simulation is a data.frame containing
the weight and numbers of praline for each box simulated
}
\examples{
# dati <- dati <- rnorm(100, 10.1, 0.5)
# b <- simulateMH(dati, 152, 150, 15)
}

