% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulatePack}
\alias{simulatePack}
\title{simulatePack}
\usage{
simulatePack(source, composition, declared, plotce = F, main = "Pesi medi",
  step = 1e+05)
}
\arguments{
\item{source}{a data frame with the weight of the single praline}

\item{composition}{a data frame with the composition of the pack}

\item{declared}{declared weight}

\item{plotce}{boolean, T if you want to plot T1 and T2, otherwise F}

\item{main}{optional main title}

\item{step}{simulation steps, number of pack to be simulated}
}
\value{
data list
}
\description{
Simulate the weight of a mixed box containing different product
using the weight of all the single product
}
\examples{
# data(source, composition)
# simulatePack(source, composition, 136)
}

