% Generated by roxygen2 (4.0.2): do not edit by hand
\name{source}
\alias{source}
\title{Weight of single praline in Thorntons Continental}
\format{A data frame with 100 rows and 10 variables:
	\describe{
			\item{Alpini}{numbers of Alpine praline present in the box}
			\item{VanillaT}{numbers of Vanilla Truffle praline present in the box}
			\item{Sicilian}{numbers of Sicilian Lemon praline present in the box}
			\item{Apricot}{numbers of Apricot Danish praline present in the box}
			\item{Coffee}{numbers of Coffee praline present in the box}
			\item{Hslice}{numbers of Hazelnut Slice Lemon praline present in the box}
			\item{Milano}{numbers of Milano praline present in the box}
			\item{Mix}{numbers of Mix praline present in the box}
			\item{Orange}{numbers of Orange praline present in the box}
			\item{Salted}{numbers of Salted Caramel praline present in the box}
	}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
source(file, local = FALSE, echo = verbose, print.eval = echo, exprs,
  spaced = use_file, verbose = getOption("verbose"),
  prompt.echo = getOption("prompt"), max.deparse.length = 150,
  width.cutoff = 60L, deparseCtrl = "showAttributes", chdir = FALSE,
  encoding = getOption("encoding"), continue.echo = getOption("continue"),
  skip.echo = 0, keep.source = getOption("keep.source"))
}
\description{
A dataset containing the weight of each praline present in the box
	the column name indicate the type of praline, in the following row
	there is the weigth of a sample of 100 single praline
}

