% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stackFB}
\alias{stackFB}
\title{stackFB}
\usage{
stackFB(dati, cluster = "plant", filter = NULL,
  main = "Product fat-bloom", new = "month", tabOnGraph = T)
}
\arguments{
\item{cluster}{field used to clusterize data}

\item{filter}{data filter}

\item{main}{main title}

\item{new}{field useed to change graph}

\item{tabOnGraph}{boolean, if T print kruskal test result above graphics}

\item{data}{data frame of data}
}
\value{
elaborated data frame
}
\description{
Simulate chocolate crystallization using Imogen Foubert model
  of crystallization kinetics of cocoa butter
}
\examples{
#tempering(BC=0.42)
}

