% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{tecTools-package}
\alias{tecTools-package}
\title{tecTool: A collection of functions for Food Technologists}
\description{
The function present in this package may help in various issues:
 1. Simulate an Enrobing machine
 2. Simulate a tempering
 3. Simulate  a pack weight
 4. Analyze a Collapse Test
}
\references{
\url{https://en.wikipedia.org/wiki/Psychrometrics}
}

