% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tempering}
\alias{tempering}
\title{tempering}
\usage{
tempering(BC = 0.42, t0 = 35, f0 = 1, a = 100, n = 3, k = 5)

tempering(BC = 0.42, t0 = 35, f0 = 1, a = 100, n = 3, k = 5)
}
\arguments{
\item{BC}{cocoa butter contents (0-1)}

\item{t0}{initial temperature of simulation (cooling) - 35 C}

\item{f0}{initially present amount of crystals [J/g or \% solid fat] - 1\%}

\item{a}{value for f as t approaches infinity [J/g or \% solid fat] - 100\%}

\item{n}{order of reverse reaction, linked to asymmetry of curve [-] - 3}

\item{k}{rate constant [h-1] - 5}

\item{BC}{cocoa butter contents (0-1)}

\item{t0}{initial temperature of simulation (cooling) - 35 C}

\item{f0}{initially present amount of crystals [J/g or \% solid fat] - 1\%}

\item{a}{value for f as t approaches infinity [J/g or \% solid fat] - 100\%}

\item{n}{order of reverse reaction, linked to asymmetry of curve [-] - 3}

\item{k}{rate constant [h-1] - 5}
}
\value{
temperature list

temperature list
}
\description{
Simulate chocolate crystallization using Imogen Foubert model
  of crystallization kinetics of cocoa butter

Simulate chocolate crystallization using Imogen Foubert model
  of crystallization kinetics of cocoa butter
}
\examples{
#tempering(BC=0.42)
#tempering(BC=0.42)
}

